Fixes edge hang position and double jump bugs
Players would not snap immediately to the correct position when hanging
from a ledge. This was fixed by setting position manually for the frame
while simultaneously instructing the rigidbody to move to the same
position. Addition of a double jumping animation revealed that it was
possible to air jump from the first frame after touching the ground,
allowing unlimited air jumps consecutively. Adding an field to the
JumpStartState to explicitly designate air jumps fixed the problem.