Adds animations and DirectInput detection
DirectInput controllers are now recognized if no xInput controllers
press A during a given frame. So far this has failed once so it will
require later debugging. Edge grabbing was slightly reworked due to a
bug noticed while recovering to a ledge in which the character grabbed and
remained above the ledge before dropping down and grabbing it again.
Dodge and block animations were added and a separate state for entering
into block was added. Transitions were added between block and dodge
states and external states. xInput controls were reworked between the
triggers and shoulder buttons for ergonomic reasons.