Fixes bugs relating to blocking and combat
A bug was detected causing sprite shaking to go out of control when forced
to change state while shaking. This was fixed by moving sprite shake code
to the player controller (possibly unnecessary) and forcing sprite shaking
to stop on state exit.
Players could also dodge without holding block onthe first frame of
BlockStart. This was fixed by transitioning back to idle from dodge
instead of to BlockStart.
CameraControl had a bug in which if no players were in bounds a list out
of bounds error would be thrown since the program was trying to find the
max of an empty list. Checking if the list was empty first fixed this.
Finally, block resistance was reworked so that smash moves can knock a
player out of it, and stun times were heavily increased.