Adds character select with a second character
Players can now choose between a fast falling character and a floaty
character, although the floaty character has not been balanced at all
and still uses all the moves and animations of the fast faller. This
will in the future be replaced by a real character however so animations
will not be balanced on the test floaty character.