Fixes the AI and makes it much smarter
Many bugs have been fixed with the GhostAI and it works decently well now.
The hard-coded AI has some ability to recover from offstage now, although
theres still some issues with the AIBehaviours properly executing
movements. I suspect this might have to do with the part of the frame
during which the AIBehaviours Process function is being called on.