Implements DI, fixes dodging, control fixes
Directional influence has finally been implemented. There was an issue
with dodging in which if dodging forward on the first frame after a
forward dodge, a backward dodge would be done instead. Not allowing
dodging until after block start fixed this issue, although this will
either have to be looked at later, or block start will have to be very
fast. Functions for AI learning were implemented on keyboard and
DirectInput controllers to avoid NotImplementedExceptions, although these
will have to be looked at and fixed later, particularly with regards to
ButtonHoldState.